Update dependencies
This updates the dependencies based on the builtin list in mkosi's
fedora default tools tree package list.

We get rid of a few old dependencies that aren't relevant anymore
such as veritysetup and binutils as well. We drop the
python3-cryptography dependency as it's a dependency of ukify, not
mkosi itself.

While the various filesystem creation tools are invoked by
systemd-repart and not mkosi itself, we still opt to keep them as
recommends in mkosi to avoid having to add them as recommends to
systemd itself.

We also get rid of the EPEL 7 macros since v15 will never be built
for EPEL 7 anyway as its a huge breaking change and they don't like
those in EPEL.

The list of dependencies is also reordered and cleaned up a bit.